[MOV] rename js_type_files
Just renaming and adapting imports. No actual code change.
