+++ server/src/core/js_module_scope.rs
