+++ server/src/core/mod.rs
@@ -12,7 +12,7 @@ pub mod file_mgr;
 pub mod import_resolver;
 pub mod js_arch_builder;
 pub mod js_import_graph;
-pub mod js_type_files;
+pub mod js_module_scope;
 pub mod js_utils;
 pub mod js_validator;
 pub mod model;
